Recreates the example project with React Native 0.58.1
