--- example/android/.settings/org.eclipse.buildship.core.prefs
@@ -1,3 +1,0 @@
-#Sat Jun 24 10:56:21 IST 2017
-connection.gradle.distribution=GRADLE_DISTRIBUTION(WRAPPER)
-connection.project.dir=
