--- example/android/app/.settings/org.eclipse.buildship.core.prefs
@@ -1,2 +1,0 @@
-#Sat Jun 24 10:56:21 IST 2017
-connection.project.dir=..
