+++ example/babel.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  presets: ["module:metro-react-native-babel-preset"]
+}
