--- example/index.android.js
@@ -1,1 +1,0 @@
-import "./app/app";
