--- example/index0.ios.js
@@ -1,1 +1,0 @@
-import "./app/app";
